feat: refactored adminservice file
