+++ server/src/service-layer/controllers/UserController.ts
@@ -75,3 +75,4 @@ export class UserController extends Controller {
     return new UserService().updateUser(userId, updates)
   }
 }
+
